5th day - Selected Counter Visual, Singleton Pattern
edit a prefab for it to have a version of the counter when it is selected.

how to update the UI: notify the code with an event handler ( that takes in a generic)  that there is a selected counter, then with an events args specifying the Selected counter.
in the SelectedCounterVisual's script the singleton pattern is used.  First, an instance property is created in the player script. That instance is then accessed in the SelectCounterVisual.