7th day - Container counter
Used inheritance  by creating a Base Counter class  for all the counters, since they all are very similar.

For the Selected counter visual. created an object array for the selected visual and a script to populate it.

created a a container counter visuals to handle the animation when the player picks an object.